Move _log_user_email to users component as it will be accessed by several components.
